Fix an update bug in which fixed-range plots were not being re-drawn. Don't create a datasource when validating it.
CCBUG: 215931



git-svn-id: svn://anonsvn.kde.org/home/kde/branches/work/kst/portto4/kst@1062534 283d02a7-25f6-0310-bc7c-ecb5cbfe19da

+++ src/libkstapp/applicationsettings.cpp
@@ -114,7 +114,7 @@ int ApplicationSettings::referenceViewWidth() const {
 #ifdef Q_WS_X11
     dpiScale = QX11Info::appDpiX();
 #endif
-  return (_refViewWidth * dpiScale * 0.3937008);
+  return int(_refViewWidth * dpiScale * 0.3937008);
 }
 
 
@@ -135,7 +135,7 @@ int ApplicationSettings::referenceViewHeight() const {
 #ifdef Q_WS_X11
     dpiScale = QX11Info::appDpiY();
 #endif
-  return (_refViewHeight * dpiScale * 0.3937008);
+  return int(_refViewHeight * dpiScale * 0.3937008);
 }
